requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.5
BeneDict==0.3.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coverage==5.4
ftfy==5.9
idna==2.10
iniconfig==1.1.1
mailchecker==4.0.3
packaging==20.9
phonenumbers==8.12.18
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pyparsing==2.4.7
pyserial==3.5
pytest==6.2.2
pytest-cov==2.11.1
python-benedict==0.23.2
python-dateutil==2.8.1
python-fsutil==0.4.0
python-slugify==4.0.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.3
wcwidth==0.2.5
xmltodict==0.12.0
